# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "screen-service API",
        "title": "screen-service API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/swagger/screen-service",
    "paths": {
        "/big-screen/delete": {
            "delete": {
                "description": "删除实例",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Big_screen"
                ],
                "summary": "删除实例",
                "parameters": [
                    {
                        "type": "string",
                        "description": "实例id",
                        "name": "ids",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "实例",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.ProjectItem"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "错误code和错误信息",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
        "/project/create": {
            "post": {
                "description": "保存实例",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Big_screen"
                ],
                "summary": "保存实例",
                "parameters": [
                    {
                        "description": "实例全部信息",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "实例",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Big_screen"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "错误code和错误信息",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/project/edit": {
            "post": {
                "description": "保存实例",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Big_screen"
                ],
                "summary": "保存实例",
                "parameters": [
                    {
                        "description": "实例全部信息",
                        "name": "item",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "实例",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Big_screen"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "错误code和错误信息",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/project/getData": {
            "get": {
                "description": "查询数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Big_screen"
                ],
                "summary": "查询数据",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "projectId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "实例的数组",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/common.Page"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "items": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/entity.ProjectItem"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "错误code和错误信息",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/project/list": {
            "get": {
                "description": "查询所有实例",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Big_screen"
                ],
                "summary": "查询所有实例",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "current page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page size",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "实例的数组",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/common.Page"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "items": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/entity.ProjectItem"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "错误code和错误信息",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/project/publish": {
            "put": {
                "description": "保存大屏内容",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Big_screen"
                ],
                "summary": "保存大屏内容",
                "parameters": [
                    {
                        "description": "data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "实例",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Big_screen"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "错误code和错误信息",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/project/save/data": {
            "post": {
                "description": "保存大屏内容",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Big_screen"
                ],
                "summary": "保存大屏内容",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "projectId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "content",
                        "name": "content",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "实例",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Big_screen"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "错误code和错误信息",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
        "/project/upload": {
            "post": {
                "description": "保存大屏封面文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Big_screen"
                ],
                "summary": "保存大屏封面文件",
                "parameters": [
                    {
                        "type": "file",
                        "description": "file",
                        "name": "object",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "实例",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Big_screen"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "错误code和错误信息",
                        "schema": {
                            "$ref": "#/definitions/common.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.Page": {
            "type": "object",
            "properties": {
                "items": {},
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "common.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "msg": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "entity.ProjectItem": {
            "type": "object",
            "properties": {
                "createTime": {
                    "type": "string"
                },
                "createUserId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "indexImage": {
                    "type": "string"
                },
                "isDelete": {
                    "type": "integer"
                },
                "projectName": {
                    "type": "string"
                },
                "remarks": {
                    "type": "string"
                },
                "state": {
                    "type": "integer"
                }
            }
        },
...
        "model.Big_screen": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "内容",
                    "type": "string"
                },
                "created_by": {
                    "description": "创建人",
                    "type": "string"
                },
                "created_time": {
                    "description": "创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "唯一标识",
                    "type": "string"
                },
                "index_image": {
                    "description": "首页图片",
                    "type": "string"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "remarks": {
                    "description": "备注",
                    "type": "string"
                },
                "state": {
                    "description": "状态（-1未发布，1发布）",
                    "type": "integer"
                },
                "updated_by": {
                    "description": "更新人",
                    "type": "string"
                },
                "updated_time": {
                    "description": "更新时间",
                    "type": "string"
                }
            }
        },
...
    }
}